requirements.txt
langchain
PyPDF
python-dotenv
streamlit
faiss-cpu
streamlit-extras